personal finance calculator

+++ python_introduction/finance_calculator.py
@@ -1,7 +1,7 @@
 #Use user input, variables, and arithmetic operations to calculate and provide feedback on a user’s monthly savings 
 # and potential future savings without applying conditional statements.
-income = int(input("Enter your monthly income: "))
-expenses = int(input ("Enter your total monthly expenses: "))
+income = float(input("Enter your monthly income: "))
+expenses = float(input ("Enter your total monthly expenses: "))
 monthly_savings = income - expenses
 annual_savings = monthly_savings * 12 + (monthly_savings * 12 * 0.05) 
 print ("Your monthly savings are $" + str(monthly_savings))
